+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/base_lithium_ion_tests.py
@@ -15,12 +15,12 @@ class BaseIntegrationTestLithiumIon:
     def test_basic_processing(self):
         options = {}
         # use Ecker parameters for nonlinear diffusion
-        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Ecker2015)
+        param = pybamm.ParameterValues("Ecker2015")
         self.run_basic_processing_test(options, parameter_values=param)
 
     def test_sensitivities(self):
         model = self.model()
-        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Ecker2015)
+        param = pybamm.ParameterValues("Ecker2015")
         modeltest = tests.StandardModelTest(model, parameter_values=param)
         modeltest.test_sensitivities(
             "Current function [A]",
@@ -29,32 +29,14 @@ class BaseIntegrationTestLithiumIon:
 
     def test_basic_processing_1plus1D(self):
         options = {"current collector": "potential pair", "dimensionality": 1}
-        var = pybamm.standard_spatial_vars
-        var_pts = {
-            var.x_n: 5,
-            var.x_s: 5,
-            var.x_p: 5,
-            var.r_n: 5,
-            var.r_p: 5,
-            var.y: 5,
-            var.z: 5,
-        }
+        var_pts = {"x_n": 5, "x_s": 5, "x_p": 5, "r_n": 5, "r_p": 5, "y": 5, "z": 5}
         model = self.model(options)
         modeltest = tests.StandardModelTest(model, var_pts=var_pts)
         modeltest.test_all(skip_output_tests=True)
 
     def test_basic_processing_2plus1D(self):
         options = {"current collector": "potential pair", "dimensionality": 2}
-        var = pybamm.standard_spatial_vars
-        var_pts = {
-            var.x_n: 5,
-            var.x_s: 5,
-            var.x_p: 5,
-            var.r_n: 5,
-            var.r_p: 5,
-            var.y: 5,
-            var.z: 5,
-        }
+        var_pts = {"x_n": 5, "x_s": 5, "x_p": 5, "r_n": 5, "r_p": 5, "y": 5, "z": 5}
         model = self.model(options)
         modeltest = tests.StandardModelTest(model, var_pts=var_pts)
         modeltest.test_all(skip_output_tests=True)
@@ -81,17 +63,13 @@ class BaseIntegrationTestLithiumIon:
 
     def test_charge(self):
         options = {"thermal": "isothermal"}
-        parameter_values = pybamm.ParameterValues(
-            chemistry=pybamm.parameter_sets.Marquis2019
-        )
+        parameter_values = pybamm.ParameterValues("Marquis2019")
         parameter_values.update({"Current function [A]": -1})
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_zero_current(self):
         options = {"thermal": "isothermal"}
-        parameter_values = pybamm.ParameterValues(
-            chemistry=pybamm.parameter_sets.Marquis2019
-        )
+        parameter_values = pybamm.ParameterValues("Marquis2019")
         parameter_values.update({"Current function [A]": 0})
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
@@ -132,7 +110,7 @@ class BaseIntegrationTestLithiumIon:
             "lithium plating": "irreversible",
             "lithium plating porosity change": "true",
         }
-        param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Chen2020_plating)
+        param = pybamm.ParameterValues("Chen2020_plating")
         self.run_basic_processing_test(options, parameter_values=param)
 
     def test_reaction_limited(self):
@@ -157,42 +135,35 @@ class BaseIntegrationTestLithiumIon:
 
     def test_loss_active_material_stress_negative(self):
         options = {"loss of active material": ("none", "stress-driven")}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_loss_active_material_stress_positive(self):
         options = {"loss of active material": ("stress-driven", "none")}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_loss_active_material_stress_both(self):
         options = {"loss of active material": "stress-driven"}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_negative_cracking(self):
         options = {"particle mechanics": ("swelling and cracking", "none")}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_positive_cracking(self):
         options = {"particle mechanics": ("none", "swelling and cracking")}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_both_cracking(self):
         options = {"particle mechanics": "swelling and cracking"}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)
 
     def test_both_swelling_only(self):
         options = {"particle mechanics": "swelling only"}
-        chemistry = pybamm.parameter_sets.Ai2020
-        parameter_values = pybamm.ParameterValues(chemistry=chemistry)
+        parameter_values = pybamm.ParameterValues("Ai2020")
         self.run_basic_processing_test(options, parameter_values=parameter_values)